+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update sample labs for Windows Server 2019 to use latest
+  evaluation ISO download URIs and edition names.
+- Removed sample `samples\Sample_WS2019_NanoDomain.xml` because it
+  is not valid for Windows Server 2019.
+
 ## [1.1.0] - 2020-08-30
 
 ### Changed